+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.12.3] - 2024-09-17
+
 ### Fixed
 
 - Handle nullable parameters to API clients better. Only send query parameters if we have a value
@@ -241,7 +243,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.2...HEAD
+[unreleased]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.3...HEAD
+[0.12.3]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.3...v0.12.2
 [0.12.2]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.2...v0.12.1
 [0.12.1]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.1...v0.12.0
 [0.12.0]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.0...v0.11.0